+++ lib/widgetbook_app.directories.g.dart
@@ -11,32 +11,33 @@
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'package:autogram/ui/app_theme.dart' as _i3;
 import 'package:autogram/ui/assets.dart' as _i4;
-import 'package:autogram/ui/fragment/show_web_page_fragment.dart' as _i20;
-import 'package:autogram/ui/screens/about_screen.dart' as _i23;
-import 'package:autogram/ui/screens/main_menu_screen.dart' as _i24;
+import 'package:autogram/ui/fragment/show_web_page_fragment.dart' as _i21;
+import 'package:autogram/ui/screens/about_screen.dart' as _i24;
+import 'package:autogram/ui/screens/main_menu_screen.dart' as _i25;
 import 'package:autogram/ui/screens/main_screen.dart' as _i2;
 import 'package:autogram/ui/screens/onboarding_accept_terms_of_service_screen.dart'
-    as _i25;
-import 'package:autogram/ui/screens/onboarding_finished_screen.dart' as _i26;
+    as _i26;
+import 'package:autogram/ui/screens/onboarding_finished_screen.dart' as _i27;
 import 'package:autogram/ui/screens/onboarding_select_signing_certificate_screen.dart'
-    as _i27;
-import 'package:autogram/ui/screens/open_document_screen.dart' as _i28;
-import 'package:autogram/ui/screens/paired_device_list_screen.dart' as _i29;
+    as _i28;
+import 'package:autogram/ui/screens/open_document_screen.dart' as _i29;
+import 'package:autogram/ui/screens/paired_device_list_screen.dart' as _i30;
 import 'package:autogram/ui/screens/present_signed_document_screen.dart'
-    as _i30;
-import 'package:autogram/ui/screens/preview_document_screen.dart' as _i31;
-import 'package:autogram/ui/screens/select_certificate_screen.dart' as _i32;
-import 'package:autogram/ui/screens/settings_screen.dart' as _i33;
-import 'package:autogram/ui/screens/show_terms_of_service_screen.dart' as _i34;
-import 'package:autogram/ui/screens/sign_document_screen.dart' as _i35;
+    as _i31;
+import 'package:autogram/ui/screens/preview_document_screen.dart' as _i32;
+import 'package:autogram/ui/screens/qr_code_scanner_screen.dart' as _i19;
+import 'package:autogram/ui/screens/select_certificate_screen.dart' as _i33;
+import 'package:autogram/ui/screens/settings_screen.dart' as _i34;
+import 'package:autogram/ui/screens/show_terms_of_service_screen.dart' as _i35;
+import 'package:autogram/ui/screens/sign_document_screen.dart' as _i36;
 import 'package:autogram/ui/screens/start_remote_document_signing_screen.dart'
-    as _i36;
+    as _i37;
 import 'package:autogram/ui/widgets/app_version_text.dart' as _i9;
 import 'package:autogram/ui/widgets/autogram_logo.dart' as _i5;
 import 'package:autogram/ui/widgets/buttons.dart' as _i6;
-import 'package:autogram/ui/widgets/certificate_picker.dart' as _i21;
+import 'package:autogram/ui/widgets/certificate_picker.dart' as _i22;
 import 'package:autogram/ui/widgets/close_button.dart' as _i7;
-import 'package:autogram/ui/widgets/dialogs.dart' as _i19;
+import 'package:autogram/ui/widgets/dialogs.dart' as _i20;
 import 'package:autogram/ui/widgets/document_visualization.dart' as _i10;
 import 'package:autogram/ui/widgets/error_content.dart' as _i11;
 import 'package:autogram/ui/widgets/html_preview.dart' as _i12;
@@ -46,7 +47,7 @@ import 'package:autogram/ui/widgets/option_picker.dart' as _i14;
 import 'package:autogram/ui/widgets/preference_tile.dart' as _i15;
 import 'package:autogram/ui/widgets/result_view.dart' as _i16;
 import 'package:autogram/ui/widgets/retry_view.dart' as _i17;
-import 'package:autogram/ui/widgets/signature_type_picker.dart' as _i22;
+import 'package:autogram/ui/widgets/signature_type_picker.dart' as _i23;
 import 'package:autogram/ui/widgets/step_indicator.dart' as _i18;
 import 'package:widgetbook/widgetbook.dart' as _i1;
 
@@ -231,6 +232,13 @@ final directories = <_i1.WidgetbookNode>[
           builder: _i18.previewStepIndicator,
         ),
       ),
+      _i1.WidgetbookLeafComponent(
+        name: '_ViewFinder',
+        useCase: _i1.WidgetbookUseCase(
+          name: '',
+          builder: _i19.previewViewFinder,
+        ),
+      ),
     ],
   ),
   _i1.WidgetbookCategory(
@@ -240,7 +248,7 @@ final directories = <_i1.WidgetbookNode>[
         name: 'BottomSheet',
         useCase: _i1.WidgetbookUseCase(
           name: 'NotificationsPermissionRationale',
-          builder: _i19.previewNotificationsPermissionRationaleModal,
+          builder: _i20.previewNotificationsPermissionRationaleModal,
         ),
       )
     ],
@@ -252,7 +260,7 @@ final directories = <_i1.WidgetbookNode>[
         name: 'ShowWebPageFragment',
         useCase: _i1.WidgetbookUseCase(
           name: 'ShowWebPageFragment',
-          builder: _i20.previewShowWebPageFragment,
+          builder: _i21.previewShowWebPageFragment,
         ),
       )
     ],
@@ -264,14 +272,14 @@ final directories = <_i1.WidgetbookNode>[
         name: 'CertificatePicker',
         useCase: _i1.WidgetbookUseCase(
           name: 'CertificatePicker',
-          builder: _i21.previewCertificatePicker,
+          builder: _i22.previewCertificatePicker,
         ),
       ),
       _i1.WidgetbookLeafComponent(
         name: 'SignatureTypePicker',
         useCase: _i1.WidgetbookUseCase(
           name: 'SignatureTypePicker',
-          builder: _i22.previewSignatureTypePicker,
+          builder: _i23.previewSignatureTypePicker,
         ),
       ),
     ],
@@ -283,14 +291,14 @@ final directories = <_i1.WidgetbookNode>[
         name: 'AboutScreen',
         useCase: _i1.WidgetbookUseCase(
           name: 'AboutScreen',
-          builder: _i23.previewAboutScreen,
+          builder: _i24.previewAboutScreen,
         ),
       ),
       _i1.WidgetbookLeafComponent(
         name: 'MainMenuScreen',
         useCase: _i1.WidgetbookUseCase(
           name: '',
-          builder: _i24.previewMainMenuScreen,
+          builder: _i25.previewMainMenuScreen,
         ),
       ),
       _i1.WidgetbookLeafComponent(
@@ -304,14 +312,14 @@ final directories = <_i1.WidgetbookNode>[
         name: 'OnboardingAcceptTermsOfServiceScreen',
         useCase: _i1.WidgetbookUseCase(
           name: 'OnboardingAcceptTermsOfServiceScreen',
-          builder: _i25.previewOnboardingAcceptTermsOfServiceScreen,
+          builder: _i26.previewOnboardingAcceptTermsOfServiceScreen,
         ),
       ),
       _i1.WidgetbookLeafComponent(
         name: 'OnboardingFinishedScreen',
         useCase: _i1.WidgetbookUseCase(
           name: 'OnboardingFinishedScreen',
-          builder: _i26.previewOnboardingFinishedScreen,
+          builder: _i27.previewOnboardingFinishedScreen,
         ),
       ),
       _i1.WidgetbookComponent(
@@ -319,20 +327,20 @@ final directories = <_i1.WidgetbookNode>[
         useCases: [
           _i1.WidgetbookUseCase(
             name: 'canceled',
-            builder: _i27.previewCanceledOnboardingSelectSigningCertificateBody,
+            builder: _i28.previewCanceledOnboardingSelectSigningCertificateBody,
           ),
           _i1.WidgetbookUseCase(
             name: 'initial',
-            builder: _i27.previewInitialOnboardingSelectSigningCertificateBody,
+            builder: _i28.previewInitialOnboardingSelectSigningCertificateBody,
           ),
           _i1.WidgetbookUseCase(
             name: 'no certificate',
             builder:
-                _i27.previewNoCertificateOnboardingSelectSigningCertificateBody,
+                _i28.previewNoCertificateOnboardingSelectSigningCertificateBody,
           ),
           _i1.WidgetbookUseCase(
             name: 'success',
-            builder: _i27.previewSuccessOnboardingSelectSigningCertificateBody,
+            builder: _i28.previewSuccessOnboardingSelectSigningCertificateBody,
           ),
         ],
       ),
@@ -341,11 +349,11 @@ final directories = <_i1.WidgetbookNode>[
         useCases: [
           _i1.WidgetbookUseCase(
             name: 'error',
-            builder: _i28.previewErrorOpenDocumentScreen,
+            builder: _i29.previewErrorOpenDocumentScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'loading',
-            builder: _i28.previewLoadingOpenDocumentScreen,
+            builder: _i29.previewLoadingOpenDocumentScreen,
           ),
         ],
       ),
@@ -353,7 +361,7 @@ final directories = <_i1.WidgetbookNode>[
         name: 'PairedDeviceListScreen',
         useCase: _i1.WidgetbookUseCase(
           name: '',
-          builder: _i29.previewPairedDeviceListScreen,
+          builder: _i30.previewPairedDeviceListScreen,
         ),
       ),
       _i1.WidgetbookComponent(
@@ -361,19 +369,19 @@ final directories = <_i1.WidgetbookNode>[
         useCases: [
           _i1.WidgetbookUseCase(
             name: 'error',
-            builder: _i30.previewErrorPresentSignedDocumentScreen,
+            builder: _i31.previewErrorPresentSignedDocumentScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'initial',
-            builder: _i30.previewInitialPresentSignedDocumentScreen,
+            builder: _i31.previewInitialPresentSignedDocumentScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'loading',
-            builder: _i30.previewLoadingPresentSignedDocumentScreen,
+            builder: _i31.previewLoadingPresentSignedDocumentScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'success',
-            builder: _i30.previewSuccessPresentSignedDocumentScreen,
+            builder: _i31.previewSuccessPresentSignedDocumentScreen,
           ),
         ],
       ),
@@ -382,40 +390,47 @@ final directories = <_i1.WidgetbookNode>[
         useCases: [
           _i1.WidgetbookUseCase(
             name: 'error',
-            builder: _i31.previewErrorPreviewDocumentScreen,
+            builder: _i32.previewErrorPreviewDocumentScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'loading',
-            builder: _i31.previewLoadingPreviewDocumentScreen,
+            builder: _i32.previewLoadingPreviewDocumentScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'success',
-            builder: _i31.previewSuccessPreviewDocumentScreen,
+            builder: _i32.previewSuccessPreviewDocumentScreen,
           ),
         ],
+      ),
+      _i1.WidgetbookLeafComponent(
+        name: 'QRCodeScannerScreen',
+        useCase: _i1.WidgetbookUseCase(
+          name: '',
+          builder: _i19.previewQRCodeScannerScreen,
+        ),
       ),
       _i1.WidgetbookComponent(
         name: 'SelectCertificateScreen',
         useCases: [
           _i1.WidgetbookUseCase(
             name: 'canceled',
-            builder: _i32.previewCanceledSelectCertificateScreen,
+            builder: _i33.previewCanceledSelectCertificateScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'error',
-            builder: _i32.previewErrorSelectCertificateScreen,
+            builder: _i33.previewErrorSelectCertificateScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'loading',
-            builder: _i32.previewLoadingSelectCertificateScreen,
+            builder: _i33.previewLoadingSelectCertificateScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'no certificate',
-            builder: _i32.previewNoCertificateSelectCertificateScreen,
+            builder: _i33.previewNoCertificateSelectCertificateScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'success',
-            builder: _i32.previewSuccessSelectCertificateScreen,
+            builder: _i33.previewSuccessSelectCertificateScreen,
           ),
         ],
       ),
@@ -423,14 +438,14 @@ final directories = <_i1.WidgetbookNode>[
         name: 'SettingsScreen',
         useCase: _i1.WidgetbookUseCase(
           name: 'SettingsScreen',
-          builder: _i33.previewSettingsScreen,
+          builder: _i34.previewSettingsScreen,
         ),
       ),
       _i1.WidgetbookLeafComponent(
         name: 'ShowTermsOfServiceScreen',
         useCase: _i1.WidgetbookUseCase(
           name: 'ShowTermsOfServiceScreen',
-          builder: _i34.previewShowTermsOfServiceScreen,
+          builder: _i35.previewShowTermsOfServiceScreen,
         ),
       ),
       _i1.WidgetbookComponent(
@@ -438,15 +453,15 @@ final directories = <_i1.WidgetbookNode>[
         useCases: [
           _i1.WidgetbookUseCase(
             name: 'error',
-            builder: _i35.previewErrorSignDocumentScreen,
+            builder: _i36.previewErrorSignDocumentScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'loading',
-            builder: _i35.previewLoadingSignDocumentScreen,
+            builder: _i36.previewLoadingSignDocumentScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'success',
-            builder: _i35.previewSuccessSignDocumentScreen,
+            builder: _i36.previewSuccessSignDocumentScreen,
           ),
         ],
       ),
@@ -454,7 +469,7 @@ final directories = <_i1.WidgetbookNode>[
         name: 'StartRemoteDocumentSigningScreen',
         useCase: _i1.WidgetbookUseCase(
           name: '',
-          builder: _i36.previewStartRemoteDocumentSigningScreen,
+          builder: _i37.previewStartRemoteDocumentSigningScreen,
         ),
       ),
     ],